+++ app/victoria-traces/main.go
@@ -129,5 +129,5 @@ See the docs at https://docs.victoriametrics.com/victoriatraces/
 
 func initSecretFlags() {
 	pushmetrics.InitSecretFlags()
-	flagutil.RegisterSecretFlag("vmalert.proxyURL")
+	vtselect.InitSecretFlags()
 }